Use data_files to copy files to site-packages
The reason we're playing whack-a-mole with these datafiles
is that MANIFEST is ignored with doing `python setup.py install`,
so we need ot use data_files or package_data. package_data for a
`lib` package ended up creating a sibling directory `lib/` in
site-packages, instead of nesting under `dash_html_components`.

The hope is that using data_files will actually copy files in under
dash_html_components consistently for local, and remote installs,
so we can call it a day.